pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "autonomy-hub"
version = "0.1.0"
description = "Planner-led portable autonomy hub for multi-repo software work."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.116.0",
  "jinja2>=3.1.6",
  "pydantic-settings>=2.10.1",
  "PyYAML>=6.0.2",
  "sqlalchemy>=2.0.43",
  "uvicorn[standard]>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1",
  "pytest>=8.4.1",
]

[project.scripts]
autonomy-hub = "autonomy_hub.main:run_dev"

[tool.hatch.build.targets.wheel]
packages = ["src/autonomy_hub"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
